query.c:   Although the documentation states that certain selectors return an array   of integers, the AUpv structure only uses long integers.  So I'm   resolving the ambiguity in the documentation by making such selectors   return an array of long integers.   This change has no impact on platforms where sizeof (int) == sizeof   (long), but it makes the query system consistent on all other   platforms. wavewrite.c:   Get rid of debugging print statements.
BKrev: 3ae697dc-9p6lF5dRhsliGkwDxAXJw

+++ libaudiofile/query.c
@@ -101,9 +101,9 @@ AUpvlist _afQueryFileFormat (int arg1, int arg2, int arg3, int arg4)
 		case AF_QUERY_IDS:
 		{
 			int	count = 0, idx;
-			int	*buffer;
+			long	*buffer;
 
-			buffer = _af_calloc(_AF_NUM_UNITS, sizeof (int));
+			buffer = _af_calloc(_AF_NUM_UNITS, sizeof (long));
 			if (buffer == NULL)
 				return AU_NULL_PVLIST;
 
@@ -174,7 +174,7 @@ AUpvlist _afQueryFileFormat (int arg1, int arg2, int arg3, int arg4)
 		case AF_QUERY_COMPRESSION_TYPES:
 		{
 			int	idx, count;
-			int	*buffer;
+			long	*buffer;
 
 			if (arg3 < 0 || arg3 >= _AF_NUM_UNITS)
 			{
@@ -194,7 +194,7 @@ AUpvlist _afQueryFileFormat (int arg1, int arg2, int arg3, int arg4)
 					if (count == 0)
 						return AU_NULL_PVLIST;
 
-					buffer = _af_calloc(count, sizeof (int));
+					buffer = _af_calloc(count, sizeof (long));
 					if (buffer == NULL)
 						return AU_NULL_PVLIST;
 
@@ -221,6 +221,7 @@ long afQueryLong (int querytype, int arg1, int arg2, int arg3, int arg4)
 	list = afQuery(querytype, arg1, arg2, arg3, arg4);
 	if (list == AU_NULL_PVLIST)
 		return -1;
+
 	AUpvgetvaltype(list, 0, &type);
 	if (type != AU_PVTYPE_LONG)
 		return -1;
@@ -238,6 +239,7 @@ double afQueryDouble (int querytype, int arg1, int arg2, int arg3, int arg4)
 	list = afQuery(querytype, arg1, arg2, arg3, arg4);
 	if (list == AU_NULL_PVLIST)
 		return -1;
+
 	AUpvgetvaltype(list, 0, &type);
 	if (type != AU_PVTYPE_DOUBLE)
 		return -1;
@@ -255,6 +257,7 @@ void *afQueryPointer (int querytype, int arg1, int arg2, int arg3, int arg4)
 	list = afQuery(querytype, arg1, arg2, arg3, arg4);
 	if (list == AU_NULL_PVLIST)
 		return NULL;
+
 	AUpvgetvaltype(list, 0, &type);
 	if (type != AU_PVTYPE_PTR)
 		return NULL;
@@ -282,14 +285,14 @@ AUpvlist _afQueryInstrumentParameter (int arg1, int arg2, int arg3, int arg4)
 		case AF_QUERY_IDS:
 		{
 			int	i, count;
-			int	*buffer;
+			long	*buffer;
 
 			if (arg2 < 0 || arg2 >= _AF_NUM_UNITS)
 				return AU_NULL_PVLIST;
 			count = _af_units[arg2].instrumentParameterCount;
 			if (count == 0)
 				return AU_NULL_PVLIST;
-			buffer = _af_calloc(count, sizeof (int));
+			buffer = _af_calloc(count, sizeof (long));
 			if (buffer == NULL)
 				return AU_NULL_PVLIST;
 			for (i=0; i<count; i++)
